% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hutchins_cols}
\alias{hutchins_cols}
\title{Function to extract hutchins colors as hex codes}
\usage{
hutchins_cols(...)
}
\arguments{
\item{...}{Character names of hutchins_colors}
}
\description{
Function to extract hutchins colors as hex codes
}
