% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hutchins_pal}
\alias{hutchins_pal}
\title{Return function to interpolate a hutchins color palette}
\usage{
hutchins_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in hutchins_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a hutchins color palette
}
