% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{hutchins_colors}
\alias{hutchins_colors}
\alias{hutchins_palettes}
\title{Urban Institute palettes}
\format{
An object of class \code{character} of length 18.

An object of class \code{list} of length 8.
}
\source{
Urban Institute Data Visualization Style Guide
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\usage{
hutchins_colors

hutchins_palettes
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Urban Institute Data Visualization Style Guide.
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\concept{hutchins palettes}
\keyword{datasets}
