% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto_install.R
\name{roboto_install}
\alias{roboto_install}
\title{Import and register Roboto font}
\usage{
roboto_install()
}
\description{
\code{roboto_install} tests to see if Roboto is imported and registered. If Roboto, isn't imported and registered, then \code{roboto_install} imports and registers Roboto with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Roboto must be installed on your computer for \code{roboto_install} to work. Roboto is the Urban Institute's main font and can be installed from \href{https://fonts.google.com/specimen/Roboto}{Google fonts}.

Test to see if Roboto is imported and registered with \code{roboto_test()}.
}
