% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto_test.R
\name{roboto_test}
\alias{roboto_test}
\title{Test for Roboto import and registration}
\usage{
roboto_test()
}
\description{
\code{roboto_test} tests to see if Roboto is imported and registered. Roboto is the Brookings Institution's main font and can be installed from \href{https://fonts.google.com/specimen/Roboto}{Google fonts}.
}
\details{
Import and register Roboto in R with roboto_install().
}
